requirements.txt
pydantic==2.8.2
langchain-openai==0.1.14
langchain-anthropic==0.1.20
python-dotenv==1.0.1
pandas==2.2.2
matplotlib==3.9.1
langchain==0.2.10
openai==1.37.1